% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus_view.R
\name{mplus.get.available.outputs}
\alias{mplus.get.available.outputs}
\title{Mplus Get Available Outputs}
\usage{
mplus.get.available.outputs(file)
}
\arguments{
\item{file}{the quoted name of an existing H5 file}
}
\value{
lists the functions available for the output
}
\description{
Lists all available functions. Similar to mplus.view.results().
}
\examples{

ex3_1 <- system.file("extdata", "ex3_1.h5", package = "mplush5")
mplus.get.available.outputs(ex3_1)

ex5_5 <- system.file("extdata", "ex5_5.h5", package = "mplush5")
mplus.get.available.outputs(ex5_5)

}
