% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus_utility.R
\name{mplus.get.dataset.attribute}
\alias{mplus.get.dataset.attribute}
\title{Mplus Get Dataset Attribute}
\usage{
mplus.get.dataset.attribute(file, datastr, attrstr)
}
\arguments{
\item{file}{the quoted name of an existing GH5 file}

\item{datastr}{the name of the group for the attribute}

\item{attrstr}{the name of the attribute}
}
\description{
Supporting function for getting attribute
}
\examples{
\dontrun{
mplus.get.dataset.attribute('ex8.1.gh5','individual_data','var_names')
}

}
