% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus_utility.R
\name{mplus.get.group.attribute}
\alias{mplus.get.group.attribute}
\title{Mplus get group attribute}
\usage{
mplus.get.group.attribute(file, groupstr, attrstr)
}
\arguments{
\item{file}{the quoted name of an existing GH5 file}

\item{groupstr}{the name of the group for the attribute}

\item{attrstr}{the name of the attribute}
}
\description{
supporting function for getting attribute
}
\examples{
\dontrun{
mplus.get.group.attribute('ex8.1.gh5','individual_data','var_names')
}

}
