% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus_print.R
\name{mplus.print.brant.wald.test}
\alias{mplus.print.brant.wald.test}
\title{Mplus Print Brant Wald Test}
\usage{
mplus.print.brant.wald.test(file)
}
\arguments{
\item{file}{the quoted name of an existing OH5 file}
}
\value{
The Brant Wald Test for Proportional Odds is printed to the screen.
}
\description{
Show the Brant Wald Test for Proportional Odds
}
\examples{
\dontrun{
mplus.print.brant.wald.test('ex.oh5')
}

}
