% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus_print.R
\name{mplus.print.confidence.intervals.in.probability.scale}
\alias{mplus.print.confidence.intervals.in.probability.scale}
\title{Mplus Print Confidence Intervals in Probability Scale}
\usage{
mplus.print.confidence.intervals.in.probability.scale(
  file,
  cred = FALSE,
  invariance = 0
)
}
\arguments{
\item{file}{the quoted name of an existing OH5 file}

\item{cred}{logical TRUE for Credibility Intervals}

\item{invariance}{0 (not invariance model, default)}
}
\value{
The confidence intervals in probability scale are printed to the screen
}
\description{
Show the results in probability scale
}
\examples{
\dontrun{
mplus.print.confidence.intervals.in.probability.scale('ex.oh5')
}

}
