% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus_print.R
\name{mplus.print.confidence.intervals}
\alias{mplus.print.confidence.intervals}
\title{Mplus Print Confidence Intervals}
\usage{
mplus.print.confidence.intervals(file, cred = FALSE, invariance = 0)
}
\arguments{
\item{file}{the quoted name of an existing OH5 file}

\item{cred}{logical TRUE for Credibility Intervals}

\item{invariance}{0 (not invariance model, default)}
}
\value{
A data frame containing the confidence intervals
}
\description{
Show the confidence intervals
}
\examples{

ex3_1 <- system.file("extdata", "ex3_1.h5", package = "mplush5")
ci <- mplus.print.confidence.intervals(ex3_1)
ci

}
