% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus_print.R
\name{mplus.print.errors.and.warnings}
\alias{mplus.print.errors.and.warnings}
\title{Mplus Print Errors and Warnings}
\usage{
mplus.print.errors.and.warnings(file, quiet = FALSE)
}
\arguments{
\item{file}{the quoted name of an existing OH5 file}

\item{quiet}{logical}
}
\value{
The errors and warning messages are printed to the screen
}
\description{
Show the errors and warnings
}
\examples{
\dontrun{
mplus.print.errors.and.warnings('ex.oh5')
}

}
