% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus_print.R
\name{mplus.print.general.results}
\alias{mplus.print.general.results}
\title{Mplus Print General Results}
\usage{
mplus.print.general.results(file, gstr, subgstr)
}
\arguments{
\item{file}{the quoted name of an existing OH5 file}

\item{gstr}{the quoted name of results group}

\item{subgstr}{the quoted name of results subgroup}
}
\value{
A data frame of the model results
}
\description{
Show the results with given group name
}
\examples{
\dontrun{
mplus.print.general.results('ex.oh5')
}

}
