% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus_print.R
\name{mplus.print.input.instructions}
\alias{mplus.print.input.instructions}
\title{Mplus Print Input Instructions}
\usage{
mplus.print.input.instructions(file, quiet = FALSE)
}
\arguments{
\item{file}{the quoted name of an existing H5 file}

\item{quiet}{logical}
}
\value{
The Mplus input instructions are printed to the screen
}
\description{
Show the input instructions.
}
\examples{

ex3_1 <- system.file("extdata", "ex3_1.h5", package = "mplush5")
mplus.print.input.instructions(ex3_1)

ex5_5 <- system.file("extdata", "ex5_5.h5", package = "mplush5")
mplus.print.input.instructions(ex5_5)

}
