% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus_print.R
\name{mplus.print.irt.parameterization}
\alias{mplus.print.irt.parameterization}
\title{Mplus Print IRT Parameterization}
\usage{
mplus.print.irt.parameterization(file, invariance = 0)
}
\arguments{
\item{file}{the quoted name of an existing H5 file}

\item{invariance}{0 (not invariance model, default)}
}
\value{
A data frame containing the IRT parameterization results
}
\description{
Show the IRT parameterization
}
\examples{

ex5_5 <- system.file("extdata", "ex5_5.h5", package = "mplush5")
irt_results <- mplus.print.irt.parameterization(ex5_5)
irt_results

}
