% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus_print.R
\name{mplus.print.latent.class.indicator.odds.ratios}
\alias{mplus.print.latent.class.indicator.odds.ratios}
\title{Mplus Print Latent Class Indicator Odds Ratios}
\usage{
mplus.print.latent.class.indicator.odds.ratios(file, invariance = 0)
}
\arguments{
\item{file}{the quoted name of an existing OH5 file}

\item{invariance}{0 (not invariance model, default)}
}
\value{
The latent class indicator odds ratios is printed to the screen
}
\description{
Show the latent class indicator odds ratios
}
\examples{
\dontrun{
mplus.print.latent.class.indicator.odds.ratios('ex.oh5')
}

}
