% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus_print.R
\name{mplus.print.logistic.regressions.odds.ratio.results}
\alias{mplus.print.logistic.regressions.odds.ratio.results}
\title{Mplus Print Logistic Regression Odds Ratio Results}
\usage{
mplus.print.logistic.regressions.odds.ratio.results(file, getdf = FALSE)
}
\arguments{
\item{file}{the quoted name of an existing OH5 file}

\item{getdf}{logical}
}
\value{
The logistic regression odds ratio results is printed to the screen
}
\description{
Show the model results
}
\examples{
\dontrun{
mplus.print.logistic.regressions.odds.ratio.results('ex.oh5')
}

}
