% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus_print.R
\name{mplus.print.model.fit.information}
\alias{mplus.print.model.fit.information}
\title{Mplus Print Model Fit Information}
\usage{
mplus.print.model.fit.information(file, invariance = 0)
}
\arguments{
\item{file}{the quoted name of an existing H5 file}

\item{invariance}{0 (not invariance model, default)}
}
\value{
A data frame containing the model fit information
}
\description{
Show the model fit information
}
\examples{

ex3_1 <- system.file("extdata", "ex3_1.h5", package = "mplush5")
mod_fit <- mplus.print.model.fit.information(ex3_1)
mod_fit

}
