% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus_print.R
\name{mplus.print.model.modification.indices}
\alias{mplus.print.model.modification.indices}
\title{Mplus Print Model Modification Indices}
\usage{
mplus.print.model.modification.indices(file, getdf = FALSE)
}
\arguments{
\item{file}{the quoted name of an existing H5 file}

\item{getdf}{logical}
}
\value{
A data frame containing the modification indices
}
\description{
Show the model modification indices
}
\examples{

ex5_5_wlsmv <- system.file("extdata", "ex5_5_wlsmv.h5", package = "mplush5")
mi <- mplus.print.model.modification.indices(ex5_5_wlsmv)
mi

}
