% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus_print.R
\name{mplus.print.model.results}
\alias{mplus.print.model.results}
\title{Mplus Print Model Results}
\usage{
mplus.print.model.results(file, invariance = 0)
}
\arguments{
\item{file}{the quoted name of an existing H5 file}

\item{invariance}{model invariance

  * `0`: (the default): not invariance model

  * `1`: configural invariance

  * `2`: metric invariance

  * `3`: scalar invariance}
}
\value{
A data frame containing the model results
}
\description{
Show the model results
}
\examples{

ex3_1 <- system.file("extdata", "ex3_1.h5", package = "mplush5")
results <- mplus.print.model.results(ex3_1)
results
ex5_5 <- system.file("extdata", "ex5_5.h5", package = "mplush5")
results <- mplus.print.model.results(ex5_5)
results

}
