% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus_print.R
\name{mplus.print.results.in.probability.scale}
\alias{mplus.print.results.in.probability.scale}
\title{Mplus Print Results in Probability Scale}
\usage{
mplus.print.results.in.probability.scale(file, invariance = 0)
}
\arguments{
\item{file}{the quoted name of an existing H5 file}

\item{invariance}{0 (not invariance model, default)}
}
\value{
A data frame containing the results in the probability scale
}
\description{
Show the results in probability scale
}
\examples{

ex5_5 <- system.file("extdata", "ex5_5.h5", package = "mplush5")
prob_results <- mplus.print.results.in.probability.scale(ex5_5)
prob_results

}
