% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus_print.R
\name{mplus.print.standardized.confidence.intervals}
\alias{mplus.print.standardized.confidence.intervals}
\title{Mplus Print Standardized Confidence Intervals}
\usage{
mplus.print.standardized.confidence.intervals(file, stype, cred = FALSE)
}
\arguments{
\item{file}{the quoted name of an existing H5 file}

\item{stype}{the standardized type ['stdyx','stdy','std']}

\item{cred}{logical TRUE for Credibility Intervals}
}
\value{
A data frame containing the standardized confidence intervals
}
\description{
Show the standardized confidence intervals
}
\examples{

ex3_1 <- system.file("extdata", "ex3_1.h5", package = "mplush5")
ci_stdyx <- mplus.print.standardized.confidence.intervals(ex3_1, "stdyx")
ci_stdyx

}
