% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus_print.R
\name{mplus.print.standardized.model.results}
\alias{mplus.print.standardized.model.results}
\title{Mplus Print Standardized Model Results}
\usage{
mplus.print.standardized.model.results(file, stype, cluster = FALSE)
}
\arguments{
\item{file}{the quoted name of an existing H5 file}

\item{stype}{the standardized type ['stdyx','stdy','std']}

\item{cluster}{logical}
}
\value{
A data frame containing the standardized results
}
\description{
Show the standardized model results
}
\examples{

ex3_1 <- system.file("extdata", "ex3_1.h5", package = "mplush5")
results_stdyx <- mplus.print.standardized.model.results(ex3_1, "stdyx")
results_stdyx

}
