% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus_print.R
\name{mplus.print.summary.analysis}
\alias{mplus.print.summary.analysis}
\title{Mplus Print Summary Analysis}
\usage{
mplus.print.summary.analysis(file, section = "all", quiet = FALSE)
}
\arguments{
\item{file}{the quoted name of an existing H5 file}

\item{section}{???}

\item{quiet}{logical}
}
\value{
The summary analysis is printed to the screen
}
\description{
Show the summary analysis
}
\examples{

ex3_1 <- system.file("extdata", "ex3_1.h5", package = "mplush5")
mplus.print.summary.analysis(ex3_1)

}
