% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus_view.R
\name{mplus.view.results}
\alias{mplus.view.results}
\title{Mplus View Results}
\usage{
mplus.view.results(file, quiet = FALSE)
}
\arguments{
\item{file}{the quoted name of an existing H5 file}

\item{quiet}{logical}
}
\value{
lists the functions available for the output
}
\description{
Loads the file and lists all available functions
}
\examples{

ex3_1 <- system.file("extdata", "ex3_1.h5", package = "mplush5")
mplus.view.results(ex3_1)


}
