% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sections}
\alias{sections}
\title{results_sections.csv}
\format{
A data frame with 53 rows and 3 columns:
\describe{
  \item{section_name}{section_name}
  \item{function_name}{function_name}
  \item{arguments}{arguments}
}
}
\source{
<https://www.statmodel.com/mplus-R/H5Results.zip>
}
\usage{
sections
}
\description{
A csv file from statsmodel.com that is needed for the functions to work
}
\keyword{datasets}
