% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{summary.data}
\alias{summary.data}
\title{summary_analysis.csv}
\format{
A data frame with 98 rows and 4 columns:
\describe{
  \item{description}{description}
  \item{storagetype}{storagetype}
  \item{datatype}{datatype}
  \item{section}{section}
}
}
\source{
<https://www.statmodel.com/mplus-R/H5Results.zip>
}
\usage{
summary.data
}
\description{
A csv file from statsmodel.com that is needed for the functions to work
}
\keyword{datasets}
